+++ tests/Mock/PayPal/Client/_fixtures/V2/GetOrderAuthorization.php
@@ -16,7 +16,7 @@ class GetOrderAuthorization
         return [
             'id' => self::ID,
             'intent' => 'AUTHORIZE',
-            'status' => 'APPROVED',
+            'status' => 'COMPLETED',
             'purchase_units' => [
                 0 => [
                     'reference_id' => 'default',