+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # REPLACE_GLOBALLY_WITH_NEXT_VERSION
 - PPI-246 - Added option for advanced logging
+- PPI-251 - Fix authorized transaction state on delayed payment collection
 
 # 3.0.3
 - PPI-20 - Fix Webhooks when payment status is already set identically